refactor: extract runs bench helpers
